refactoring file,variable, copy and memory nodes
